requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
python-dotenv>=1.0.1
Flask>=3.0.0
twilio>=9.0.0
openpyxl>=3.1.0
